functions.py update

+++ Reportes/functions.py
@@ -129,19 +129,22 @@ class ModificacionesTablas():
         p_id_persona      = request.POST.get('p_id_persona')
 
         lista             = request.POST.getlist('accesorios')
+
         accesorios = Accesorio.objects.all()
-        for a in accesorios:
-            detalle = AccesorioReporte.objects.filter(r_id_reporte = id_report, a_id_accesorio = a)
-            if detalle:
-                detalle.delete()
-            
+        if accesorios:
+            for a in accesorios:
+                detalle = AccesorioReporte.objects.filter(r_id_reporte = id_report, a_id_accesorio = a)
+                if detalle:
+                    detalle.delete()
+            if lista:
+                ModificacionesTablas.crear_detalle(lista, reporte)
 
         img_maquinaria    = request.FILES.get('img_maquinaria')
         img_report        = request.FILES.get('img_report')
 
         reporte = Reporte.objects.get(id_reporte = id_report)
         usuario = Usuario.objects.get(p_id_persona = p_id_persona)
-        persona = Persona.objects.get(id_persona = usuario.p_id_persona.id_persona)
+        # persona = Persona.objects.get(id_persona = usuario.p_id_persona.id_persona)
 
         reporte.cliente           = cliente
         reporte.obra              = obra
@@ -157,8 +160,6 @@ class ModificacionesTablas():
         reporte.observaciones     = observaciones
         reporte.u_p_id_persona    = usuario
 
-        if lista:
-            ModificacionesTablas.crear_detalle(lista, reporte)
 
         if img_maquinaria == None:
             img_maquinaria = reporte.img_maquinaria
